// packages/api/src/types/vocadb.d.ts
export type VDBStatusType = "Draft" | "Finished" | "Approved" | "Locked";
export type VDBAlbumDiscType =
  | "Unknown"
  | "Album"
  | "Single"
  | "EP"
  | "SplitAlbum"
  | "Compilation"
  | "Video"
  | "Artbook"
  | "Game"
  | "Fanmade"
  | "Instrumental"
  | "Other";
export type VDBArtistCategoryType =
  | "Nothing"
  | "Vocalist"
  | "Producer"
  | "Animator"
  | "Label"
  | "Circle"
  | "Other"
  | "Band"
  | "Illustrator"
  | "Subject";
export type VDBArtistRoleType =
  | "Default"
  | "Animator"
  | "Arranger"
  | "Composer"
  | "Distributor"
  | "Illustrator"
  | "Instrumentalist"
  | "Lyricist"
  | "Mastering"
  | "Publisher"
  | "Vocalist"
  | "VoiceManipulator"
  | "Other"
  | "Mixer"
  | "Chorus"
  | "Encoder"
  | "VocalDataProvider";
export type VDBTranslationType = "Original" | "Romanized" | "Translation";
export type VDBReleaseEventCategoryType =
  | "Unspecified"
  | "AlbumRelease"
  | "Anniversary"
  | "Club"
  | "Concert"
  | "Contest"
  | "Convention"
  | "Other";
export type VDBWebLinkCategoryType =
  | "Official"
  | "Commercial"
  | "Reference"
  | "Other";
export type VDBArtistType =
  | "Unknown"
  | "Circle"
  | "Label"
  | "Producer"
  | "Animator"
  | "Illustrator"
  | "Lyricist"
  | "Vocaloid"
  | "UTAU"
  | "CeVIO"
  | "OtherVoiceSynthesizer"
  | "OtherVocalist"
  | "OtherGroup"
  | "OtherIndividual"
  | "Utaite"
  | "Band"
  | "Vocalist"
  | "Character"
  | "SynthesizerV"
  | "CoverArtist"
  | "NEUTRINO"
  | "VoiSona"
  | "NewType"
  | "Voiceroid";
// ...
export interface ArtistContract {
  additionalNames: string;
  artistType: VDBArtistType;
  deleted: boolean;
  id: number;
  name: string;
  pictureMime: string;
  releaseDate?: string;
  status: VDBStatusType;
  version: number;
}
// ...
export interface ArtistForAlbumForApiContract {
  artist: ArtistContract;
  /** VDBArtistCategoryType[].join(", ") */
  categories: string;
  /** VDBArtistRoleType[].join(", ") */
  effectiveRoles: string;
  isSupport: boolean;
  name: string;
  /** VDBArtistRoleType[].join(", ") */
  roles: string;
}